added show code command to display whole code in a highlighted manner and moved color diff to utils
